pyproject.toml
[project]
name = "template-extraction"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.2.14",
    "python-dotenv>=1.1.0",
]
